pyproject.toml
[project]
name = "box-ai-agents-toolkit"
version = "0.0.1"
description = "A python library for building AI agents for Box"
license = "MIT"
readme = "README.md"
authors = [{ email = "devrel@box.com" }]
requires-python = ">=3.11"
dependencies = [
    "box-sdk-gen>=1.12.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.9.10",
]
